refactor: split interaction.ts into focused modules (#59)
Extract CDP interaction primitives into interaction-helpers.ts and
wait_for polling logic into wait-for.ts. interaction.ts becomes a
thin wiring layer for tool registrations.

No behavior change — public API (registerInteractionTools,
waitForPossibleNavigation) unchanged. All 508 tests pass.

// ticktockbent
