+++ src/tools/interaction.ts
@@ -1,7 +1,7 @@
 import * as fs from "node:fs/promises";
 import { z } from "zod";
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import type { Page, KeyInput } from "puppeteer";
+import type { KeyInput } from "puppeteer";
 import { CharlotteError, CharlotteErrorCode } from "../types/errors.js";
 import { logger } from "../utils/logger.js";
 import type { RegisteredTool } from "@modelcontextprotocol/sdk/server/mcp.js";
@@ -14,402 +14,22 @@ import {
   handleToolError,
   coercedBoolean,
 } from "./tool-helpers.js";
+import {
+  MODIFIER_KEY_MAP,
+  clickElementByBackendNodeId,
+  focusElementByBackendNodeId,
+  hoverElementByBackendNodeId,
+  dragElementToElement,
+  typeIntoElement,
+  selectOptionByBackendNodeId,
+  submitFormByBackendNodeId,
+  setFileInputFiles,
+  waitForPossibleNavigation,
+} from "./interaction-helpers.js";
+import { registerWaitForTools } from "./wait-for.js";
 
-/** Maps short modifier names to Puppeteer KeyInput values. */
-const MODIFIER_KEY_MAP: Record<string, KeyInput> = {
-  ctrl: "Control" as KeyInput,
-  shift: "Shift" as KeyInput,
-  alt: "Alt" as KeyInput,
-  meta: "Meta" as KeyInput,
-};
-
-/**
- * Click an element by backend node ID using CDP to get coordinates,
- * or more simply by resolving to an XPath/selector and using page.click.
- *
- * The most reliable approach: use CDP to get the element's coordinates, then click at those coords.
- */
-async function clickElementByBackendNodeId(
-  page: Page,
-  backendNodeId: number,
-  clickType: "left" | "right" | "double" = "left",
-  modifiers: Array<"ctrl" | "shift" | "alt" | "meta"> = [],
-): Promise<void> {
-  // Get the element's box model to find clickable coordinates
-  const cdpSession = await page.createCDPSession();
-  try {
-    // First, scroll the element into view
-    await cdpSession.send("DOM.scrollIntoViewIfNeeded", { backendNodeId });
-
-    // Get box model for coordinates
-    const { model } = await cdpSession.send("DOM.getBoxModel", {
-      backendNodeId,
-    });
-
-    if (!model) {
-      throw new CharlotteError(
-        CharlotteErrorCode.ELEMENT_NOT_FOUND,
-        "Element has no visible box model — it may be hidden or zero-sized.",
-        "Call charlotte:observe to check the element's state.",
-      );
-    }
-
-    // content quad: [x1,y1, x2,y2, x3,y3, x4,y4]
-    const contentQuad = model.content;
-    const centerX = (contentQuad[0] + contentQuad[2] + contentQuad[4] + contentQuad[6]) / 4;
-    const centerY = (contentQuad[1] + contentQuad[3] + contentQuad[5] + contentQuad[7]) / 4;
-
-    // Hold down modifier keys before the click
-    for (const modifier of modifiers) {
-      const modifierKey = MODIFIER_KEY_MAP[modifier];
-      await page.keyboard.down(modifierKey);
-    }
-
-    try {
-      if (clickType === "right") {
-        await page.mouse.click(centerX, centerY, { button: "right" });
-      } else if (clickType === "double") {
-        await page.mouse.click(centerX, centerY, { clickCount: 2 });
-      } else {
-        await page.mouse.click(centerX, centerY);
-      }
-    } finally {
-      // Release modifier keys in reverse order (always release even if click fails)
-      for (const modifier of [...modifiers].reverse()) {
-        const modifierKey = MODIFIER_KEY_MAP[modifier];
-        await page.keyboard.up(modifierKey);
-      }
-    }
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Wait for any navigation triggered by an action, or fall back to a brief settle pause.
- *
- * Listens for the `framenavigated` CDP event to detect if a click caused navigation.
- * If navigation is detected within `detectionWindowMs`, waits for the page load event
- * (up to `loadTimeoutMs`). If no navigation fires, returns after `settleMs`.
- *
- * Also races against dialog events — if the action triggers a JavaScript dialog
- * (alert, confirm, prompt, beforeunload), the action promise will block indefinitely.
- * This function detects that and returns early so the caller can surface `pending_dialog`.
- */
-export async function waitForPossibleNavigation(
-  page: Page,
-  action: () => Promise<void>,
-  { detectionWindowMs = 500, loadTimeoutMs = 10000, settleMs = 50 } = {},
-): Promise<void> {
-  let navigationDetected = false;
-  let dialogDetected = false;
-
-  // Listen for navigation start via page event (fires on any navigation)
-  const navigationStartPromise = new Promise<void>((resolve) => {
-    const handler = () => {
-      navigationDetected = true;
-      page.off("framenavigated", handler);
-      resolve();
-    };
-    page.on("framenavigated", handler);
-
-    // Clean up listener if no navigation fires within detection window
-    setTimeout(() => {
-      page.off("framenavigated", handler);
-      resolve();
-    }, detectionWindowMs);
-  });
-
-  // Listen for dialog (blocks the action from completing)
-  const dialogPromise = new Promise<void>((resolve) => {
-    const handler = () => {
-      dialogDetected = true;
-      page.off("dialog", handler);
-      resolve();
-    };
-    page.on("dialog", handler);
-    // Clean up on timeout — if no dialog fires, we don't need this listener
-    setTimeout(() => {
-      page.off("dialog", handler);
-      resolve();
-    }, detectionWindowMs);
-  });
-
-  // Race: action vs dialog
-  const actionPromise = action();
-  await Promise.race([
-    actionPromise.then(() => "action" as const),
-    dialogPromise.then(() => "dialog" as const),
-  ]);
-
-  if (dialogDetected) {
-    // Dialog is blocking the action. Don't await actionPromise — it will
-    // resolve later when the dialog is handled via charlotte:dialog.
-    // Guard against unhandled rejection from the fire-and-forget promise.
-    actionPromise.catch(() => {
-      logger.debug("Post-dialog action promise rejected (expected)");
-    });
-    return;
-  }
-
-  // Action completed normally — check for navigation
-  await navigationStartPromise;
-
-  if (navigationDetected) {
-    // Navigation occurred — wait for the page to finish loading
-    try {
-      await page.waitForNavigation({ waitUntil: "load", timeout: loadTimeoutMs });
-    } catch {
-      // Page may have already finished loading before we called waitForNavigation,
-      // or the load timed out. Either way, render what we have.
-      logger.debug("Post-navigation load wait ended (page may already be loaded)");
-    }
-  } else {
-    // No navigation — brief settle for in-page DOM updates
-    await new Promise((resolve) => setTimeout(resolve, settleMs));
-  }
-}
-
-/**
- * Focus an element by backend node ID using CDP.
- */
-async function focusElementByBackendNodeId(page: Page, backendNodeId: number): Promise<void> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    await cdpSession.send("DOM.focus", { backendNodeId });
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Hover over an element by backend node ID.
- */
-async function hoverElementByBackendNodeId(page: Page, backendNodeId: number): Promise<void> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    await cdpSession.send("DOM.scrollIntoViewIfNeeded", { backendNodeId });
-    const { model } = await cdpSession.send("DOM.getBoxModel", {
-      backendNodeId,
-    });
-
-    if (!model) {
-      throw new CharlotteError(
-        CharlotteErrorCode.ELEMENT_NOT_FOUND,
-        "Element has no visible box model for hover.",
-      );
-    }
-
-    const contentQuad = model.content;
-    const centerX = (contentQuad[0] + contentQuad[2] + contentQuad[4] + contentQuad[6]) / 4;
-    const centerY = (contentQuad[1] + contentQuad[3] + contentQuad[5] + contentQuad[7]) / 4;
-
-    await page.mouse.move(centerX, centerY);
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Get the center coordinates of an element by backend node ID.
- * Scrolls the element into view first.
- */
-async function getElementCenter(
-  page: Page,
-  backendNodeId: number,
-): Promise<{ x: number; y: number }> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    await cdpSession.send("DOM.scrollIntoViewIfNeeded", { backendNodeId });
-    const { model } = await cdpSession.send("DOM.getBoxModel", {
-      backendNodeId,
-    });
-
-    if (!model) {
-      throw new CharlotteError(
-        CharlotteErrorCode.ELEMENT_NOT_FOUND,
-        "Element has no visible box model — it may be hidden or zero-sized.",
-        "Call charlotte:observe to check the element's state.",
-      );
-    }
-
-    const contentQuad = model.content;
-    return {
-      x: (contentQuad[0] + contentQuad[2] + contentQuad[4] + contentQuad[6]) / 4,
-      y: (contentQuad[1] + contentQuad[3] + contentQuad[5] + contentQuad[7]) / 4,
-    };
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Drag one element to another using mouse primitives.
- * Sequence: move to source → mousedown → move to target → mouseup
- * Includes intermediate move steps and delays to ensure drag events fire reliably.
- */
-async function dragElementToElement(
-  page: Page,
-  sourceBackendNodeId: number,
-  targetBackendNodeId: number,
-): Promise<void> {
-  const sourceCenter = await getElementCenter(page, sourceBackendNodeId);
-  const targetCenter = await getElementCenter(page, targetBackendNodeId);
-
-  // Move to source and press down
-  await page.mouse.move(sourceCenter.x, sourceCenter.y);
-  await page.mouse.down();
-
-  // Intermediate move to trigger dragstart (some browsers need movement to begin a drag)
-  await page.mouse.move(
-    sourceCenter.x + (targetCenter.x - sourceCenter.x) * 0.25,
-    sourceCenter.y + (targetCenter.y - sourceCenter.y) * 0.25,
-    { steps: 5 },
-  );
-  await new Promise((resolve) => setTimeout(resolve, 50));
-
-  // Move to target
-  await page.mouse.move(targetCenter.x, targetCenter.y, { steps: 10 });
-  await new Promise((resolve) => setTimeout(resolve, 50));
-
-  // Release
-  await page.mouse.up();
-}
-
-/**
- * Type text into an input element. Uses CDP to focus, optionally clears, then types via keyboard.
- */
-async function typeIntoElement(
-  page: Page,
-  backendNodeId: number,
-  text: string,
-  clearFirst: boolean,
-  pressEnter: boolean,
-): Promise<void> {
-  // Focus the element
-  await focusElementByBackendNodeId(page, backendNodeId);
-
-  if (clearFirst) {
-    // Select all text then delete — works cross-platform
-    await page.keyboard.down("Control");
-    await page.keyboard.press("a");
-    await page.keyboard.up("Control");
-    await page.keyboard.press("Backspace");
-  }
-
-  // Type the text character by character
-  await page.keyboard.type(text);
-
-  if (pressEnter) {
-    await page.keyboard.press("Enter");
-  }
-}
-
-/**
- * Select a value in a <select> element using CDP to set the value and dispatch change events.
- */
-async function selectOptionByBackendNodeId(
-  page: Page,
-  backendNodeId: number,
-  value: string,
-): Promise<void> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    // Resolve the node to get a remote object reference
-    const { object } = await cdpSession.send("DOM.resolveNode", {
-      backendNodeId,
-    });
-
-    if (!object?.objectId) {
-      throw new CharlotteError(
-        CharlotteErrorCode.ELEMENT_NOT_FOUND,
-        "Could not resolve select element.",
-      );
-    }
-
-    // Use Runtime.callFunctionOn to set the value and fire events
-    await cdpSession.send("Runtime.callFunctionOn", {
-      objectId: object.objectId,
-      functionDeclaration: `function(targetValue) {
-        const options = Array.from(this.options);
-        const matchByValue = options.find(o => o.value === targetValue);
-        const matchByText = options.find(o => o.textContent.trim() === targetValue);
-        const match = matchByValue || matchByText;
-        if (match) {
-          this.value = match.value;
-          this.dispatchEvent(new Event('input', { bubbles: true }));
-          this.dispatchEvent(new Event('change', { bubbles: true }));
-        } else {
-          throw new Error('Option "' + targetValue + '" not found');
-        }
-      }`,
-      arguments: [{ value }],
-    });
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Submit a form by backend node ID — calls form.submit() via CDP.
- */
-async function submitFormByBackendNodeId(page: Page, backendNodeId: number): Promise<void> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    const { object } = await cdpSession.send("DOM.resolveNode", {
-      backendNodeId,
-    });
-
-    if (!object?.objectId) {
-      throw new CharlotteError(
-        CharlotteErrorCode.ELEMENT_NOT_FOUND,
-        "Could not resolve form element.",
-      );
-    }
-
-    await cdpSession.send("Runtime.callFunctionOn", {
-      objectId: object.objectId,
-      functionDeclaration: `function() {
-        this.dispatchEvent(new Event('submit', { bubbles: true, cancelable: true }));
-      }`,
-    });
-  } finally {
-    await cdpSession.detach();
-  }
-}
-
-/**
- * Set files on a file input element using CDP DOM.setFileInputFiles.
- * Validates that the target element is actually an <input type="file">.
- */
-async function setFileInputFiles(
-  page: Page,
-  backendNodeId: number,
-  filePaths: string[],
-): Promise<void> {
-  const cdpSession = await page.createCDPSession();
-  try {
-    const { node } = await cdpSession.send("DOM.describeNode", { backendNodeId });
-    const isFileInput =
-      node.nodeName === "INPUT" &&
-      (node.attributes ?? []).some(
-        (attr: string, i: number, arr: string[]) => attr === "type" && arr[i + 1] === "file",
-      );
-    if (!isFileInput) {
-      throw new CharlotteError(
-        CharlotteErrorCode.SESSION_ERROR,
-        "Element is not a file input.",
-        "Use charlotte:find to locate file_input elements.",
-      );
-    }
-    await cdpSession.send("DOM.setFileInputFiles", {
-      files: filePaths,
-      backendNodeId,
-    });
-  } finally {
-    await cdpSession.detach();
-  }
-}
+// Re-export for backward compatibility (used by dialog and popup integration tests)
+export { waitForPossibleNavigation } from "./interaction-helpers.js";
 
 export function registerInteractionTools(
   server: McpServer,
@@ -1003,207 +623,9 @@ export function registerInteractionTools(
     },
   );
 
-  // ─── charlotte:wait_for ───
-  tools["charlotte:wait_for"] = server.registerTool(
-    "charlotte:wait_for",
-    {
-      description:
-        "Wait for a condition to be met on the page. Returns page representation when the condition is satisfied, or a TIMEOUT error.",
-      inputSchema: {
-        element_id: z.string().optional().describe("Wait for specific element to appear/change"),
-        state: z
-          .enum(["visible", "hidden", "enabled", "disabled", "exists", "removed"])
-          .optional()
-          .describe("Target element state to wait for"),
-        text: z.string().optional().describe("Wait for text to appear on the page"),
-        selector: z.string().optional().describe("Wait for CSS selector to match"),
-        js: z.string().optional().describe("Wait for JS expression to return truthy"),
-        timeout: z.number().optional().describe("Max wait in ms (default: 10000)"),
-      },
-    },
-    async ({ element_id, state, text, selector, js, timeout }) => {
-      try {
-        await deps.browserManager.ensureConnected();
-        const page = deps.pageManager.getActivePage();
-        const waitTimeout = timeout ?? 10000;
-
-        // Validate that at least one condition is provided
-        if (!element_id && !text && !selector && !js) {
-          throw new CharlotteError(
-            CharlotteErrorCode.SESSION_ERROR,
-            "At least one wait condition is required (element_id, text, selector, or js).",
-          );
-        }
-
-        logger.info("Waiting for condition", {
-          element_id,
-          state,
-          text,
-          selector,
-          js,
-          timeout: waitTimeout,
-        });
-
-        // Build a composite wait condition
-        const satisfied = await pollWaitForCondition(
-          deps,
-          page,
-          { element_id, state, text, selector, js },
-          waitTimeout,
-        );
-
-        if (!satisfied) {
-          const representation = await renderAfterAction(deps);
-          const timeoutError = new CharlotteError(
-            CharlotteErrorCode.TIMEOUT,
-            `Wait condition not met within ${waitTimeout}ms.`,
-            "The current page state is included in the response. Consider increasing timeout or adjusting your condition.",
-          );
-          return {
-            content: [
-              {
-                type: "text" as const,
-                text: JSON.stringify({
-                  ...timeoutError.toResponse(),
-                  page: representation,
-                }),
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        const representation = await renderAfterAction(deps);
-        return formatPageResponse(representation);
-      } catch (error: unknown) {
-        return handleToolError(error);
-      }
-    },
-  );
+  // ─── charlotte:wait_for (delegated to wait-for.ts) ───
+  const waitForTools = registerWaitForTools(server, deps);
+  Object.assign(tools, waitForTools);
 
   return tools;
-}
-
-/**
- * Poll for complex wait_for conditions that may involve element state checks.
- */
-async function pollWaitForCondition(
-  deps: ToolDependencies,
-  page: Page,
-  condition: {
-    element_id?: string;
-    state?: string;
-    text?: string;
-    selector?: string;
-    js?: string;
-  },
-  timeoutMs: number,
-): Promise<boolean> {
-  const pollInterval = 100;
-  const deadline = Date.now() + timeoutMs;
-
-  while (Date.now() < deadline) {
-    let allSatisfied = true;
-
-    // Check element_id + state condition
-    if (condition.element_id) {
-      const targetState = condition.state ?? "exists";
-      const elementSatisfied = await checkElementCondition(deps, condition.element_id, targetState);
-      if (!elementSatisfied) allSatisfied = false;
-    }
-
-    // Check text condition
-    if (allSatisfied && condition.text) {
-      const textFound = await page.evaluate((searchText) => {
-        return document.body?.innerText?.includes(searchText) ?? false;
-      }, condition.text);
-      if (!textFound) allSatisfied = false;
-    }
-
-    // Check selector condition
-    if (allSatisfied && condition.selector) {
-      const selectorMatched = await page.$(condition.selector);
-      if (!selectorMatched) allSatisfied = false;
-    }
-
-    // Check JS condition via CDP Runtime.evaluate (matches evaluate.ts pattern)
-    if (allSatisfied && condition.js) {
-      const cdpSession = await page.createCDPSession();
-      try {
-        const evalResult = await cdpSession.send("Runtime.evaluate", {
-          expression: condition.js,
-          returnByValue: true,
-          awaitPromise: true,
-          timeout: Math.max(0, deadline - Date.now()),
-        });
-        const isTruthy =
-          !evalResult.exceptionDetails && !!evalResult.result.value;
-        if (!isTruthy) allSatisfied = false;
-      } catch {
-        allSatisfied = false;
-      } finally {
-        await cdpSession.detach().catch(() => {});
-      }
-    }
-
-    if (allSatisfied) return true;
-
-    const remainingTime = deadline - Date.now();
-    if (remainingTime <= 0) break;
-
-    await new Promise((resolve) => setTimeout(resolve, Math.min(pollInterval, remainingTime)));
-  }
-
-  return false;
-}
-
-/**
- * Check if an element meets a specific state condition.
- */
-async function checkElementCondition(
-  deps: ToolDependencies,
-  elementId: string,
-  targetState: string,
-): Promise<boolean> {
-  switch (targetState) {
-    case "exists": {
-      const backendNodeId = deps.elementIdGenerator.resolveId(elementId);
-      return backendNodeId !== null;
-    }
-    case "removed": {
-      const backendNodeId = deps.elementIdGenerator.resolveId(elementId);
-      if (backendNodeId !== null) {
-        // Re-render to check if it's truly still there
-        await renderActivePage(deps, { detail: "minimal" });
-        return deps.elementIdGenerator.resolveId(elementId) === null;
-      }
-      return true;
-    }
-    case "visible":
-    case "hidden":
-    case "enabled":
-    case "disabled": {
-      // Re-render to get fresh state
-      const representation = await renderActivePage(deps, { detail: "minimal" });
-      const element = representation.interactive.find((el) => el.id === elementId);
-      if (!element) {
-        // Element doesn't exist — "hidden" and "disabled" are satisfied, others not
-        return targetState === "hidden" || targetState === "disabled";
-      }
-
-      switch (targetState) {
-        case "visible":
-          return element.state.visible === true;
-        case "hidden":
-          return element.state.visible === false;
-        case "enabled":
-          return element.state.enabled === true;
-        case "disabled":
-          return element.state.enabled === false;
-      }
-      return false;
-    }
-    default:
-      return false;
-  }
 }
